Create HOC for custom dimensions.
